+++ yt/yt/library/profiling/solomon/config.cpp
@@ -32,6 +32,28 @@ void TSolomonRegistryDynamicConfig::Register(TRegistrar registrar)
 
 ////////////////////////////////////////////////////////////////////////////////
 
+void TSolomonExporterDynamicConfig::Register(TRegistrar registrar)
+{
+    registrar.Parameter("thread_pool_size", &TThis::ThreadPoolSize)
+        .Default();
+    registrar.Parameter("thread_pool_polling_period", &TThis::ThreadPoolPollingPeriod)
+        .Default();
+}
+
+////////////////////////////////////////////////////////////////////////////////
+
+TSolomonExporterConfigPtr TSolomonExporterConfig::ApplyDynamic(
+    const TSolomonExporterDynamicConfigPtr& dynamicConfig) const
+{
+    auto result = CloneYsonStruct(MakeStrong(this));
+    NYTree::UpdateYsonStructField(result->ThreadPoolSize, dynamicConfig->ThreadPoolSize);
+    NYTree::UpdateYsonStructField(result->ThreadPoolPollingPeriod, dynamicConfig->ThreadPoolPollingPeriod);
+    result->Postprocess();
+    return result;
+}
+
+////////////////////////////////////////////////////////////////////////////////
+
 void TShardConfig::Register(TRegistrar registrar)
 {
     registrar.Parameter("filter", &TThis::Filter)